Add New Register for InstStruct

+++ include/config.h
@@ -19,7 +19,7 @@ namespace CVM
 		using DataIndexType = std::uint32_t;
 		using StackSizeType = std::uint32_t;
 		using FuncIndexType = std::uint32_t;
-		using StackOffsetType = std::uint16_t;
+		using StackOffsetType = std::int32_t;
 		using LineCountType = std::uint32_t;
 
 		constexpr bool isCheckMemorySizeOverflow = true;
